Restructured Camel forge commands into sub packages so its easier to manage
