+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/CamelAddComponentInstanceSpringCommand.java
@@ -15,6 +15,10 @@
  */
 package io.fabric8.forge.camel.commands.project;
 
+import io.fabric8.forge.camel.commands.project.completer.PackageNameCompleter;
+import io.fabric8.forge.camel.commands.project.helper.CamelCommandsHelper;
+import io.fabric8.forge.camel.commands.project.validator.ClassNameValidator;
+import io.fabric8.forge.camel.commands.project.validator.PackageNameValidator;
 import org.jboss.forge.addon.dependencies.DependencyResolver;
 import org.jboss.forge.addon.facets.constraints.FacetConstraint;
 import org.jboss.forge.addon.parser.java.facets.JavaSourceFacet;
@@ -82,7 +86,7 @@ public class CamelAddComponentInstanceSpringCommand extends AbstractCamelProject
     public boolean isEnabled(UIContext context) {
         boolean enabled = super.isEnabled(context);
         if (enabled) {
-            return CamelCommands.isSpringProject(getSelectedProject(context));
+            return CamelCommandsHelper.isSpringProject(getSelectedProject(context));
         }
         return false;
     }
@@ -93,9 +97,9 @@ public class CamelAddComponentInstanceSpringCommand extends AbstractCamelProject
         Project project = getSelectedProject(builder.getUIContext());
         JavaSourceFacet facet = project.getFacet(JavaSourceFacet.class);
 
-        componentNameFilter.setValueChoices(CamelCommands.createComponentNameValues(project));
+        componentNameFilter.setValueChoices(CamelCommandsHelper.createComponentNameValues(project));
         componentNameFilter.setDefaultValue("<all>");
-        componentName.setValueChoices(CamelCommands.createComponentNameValues(project, componentNameFilter, false));
+        componentName.setValueChoices(CamelCommandsHelper.createComponentNameValues(project, componentNameFilter, false));
 
         instanceName.setDefaultValue(new Callable<String>() {
             @Override
