+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/CamelInfoCommand.java
@@ -29,9 +29,9 @@ import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
 
-import static io.fabric8.forge.camel.commands.project.CamelCatalogHelper.componentsFromArtifact;
-import static io.fabric8.forge.camel.commands.project.CamelCatalogHelper.dataFormatsFromArtifact;
-import static io.fabric8.forge.camel.commands.project.CamelCatalogHelper.languagesFromArtifact;
+import static io.fabric8.forge.camel.commands.project.helper.CamelCatalogHelper.componentsFromArtifact;
+import static io.fabric8.forge.camel.commands.project.helper.CamelCatalogHelper.dataFormatsFromArtifact;
+import static io.fabric8.forge.camel.commands.project.helper.CamelCatalogHelper.languagesFromArtifact;
 
 public class CamelInfoCommand extends AbstractCamelProjectCommand {
