+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/CamelSetupCommand.java
@@ -20,6 +20,7 @@ import java.util.Set;
 import javax.inject.Inject;
 
 import io.fabric8.forge.camel.commands.jolokia.ConnectCommand;
+import io.fabric8.forge.camel.commands.project.helper.VersionHelper;
 import org.jboss.forge.addon.dependencies.Dependency;
 import org.jboss.forge.addon.dependencies.builder.DependencyBuilder;
 import org.jboss.forge.addon.maven.plugins.MavenPluginBuilder;
