+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/ConfigureEndpointPropertiesStep.java
@@ -23,6 +23,8 @@ import java.util.Map;
 import java.util.Set;
 import javax.inject.Inject;
 
+import io.fabric8.forge.camel.commands.project.helper.CamelCommandsHelper;
+import io.fabric8.forge.camel.commands.project.helper.CamelProjectHelper;
 import org.apache.camel.catalog.CamelCatalog;
 import org.apache.camel.catalog.DefaultCamelCatalog;
 import org.apache.camel.catalog.JSonSchemaHelper;
@@ -45,9 +47,9 @@ import org.jboss.forge.roaster.model.source.JavaClassSource;
 import org.jboss.forge.roaster.model.source.MethodSource;
 import org.jboss.forge.roaster.model.util.Strings;
 
-import static io.fabric8.forge.camel.commands.project.CamelCommands.ensureCamelArtifactIdAdded;
-import static io.fabric8.forge.camel.commands.project.CamelCommands.loadCamelComponentDetails;
-import static io.fabric8.forge.camel.commands.project.UIHelper.createUIInput;
+import static io.fabric8.forge.camel.commands.project.helper.CamelCommandsHelper.ensureCamelArtifactIdAdded;
+import static io.fabric8.forge.camel.commands.project.helper.CamelCommandsHelper.loadCamelComponentDetails;
+import static io.fabric8.forge.camel.commands.project.helper.UIHelper.createUIInput;
 
 public class ConfigureEndpointPropertiesStep extends AbstractCamelProjectCommand implements UIWizardStep {
 
@@ -89,7 +91,7 @@ public class ConfigureEndpointPropertiesStep extends AbstractCamelProjectCommand
                 String enums = propertyMap.get("enum");
 
                 if (!Strings.isNullOrEmpty(name)) {
-                    Class<Object> inputClazz = CamelCommands.loadValidInputTypes(javaType, type);
+                    Class<Object> inputClazz = CamelCommandsHelper.loadValidInputTypes(javaType, type);
                     if (inputClazz != null) {
                         if (namesAdded.add(name)) {
                             InputComponent input = createUIInput(componentFactory, name, inputClazz, required, defaultValue, enums, description);
