+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/DockerStepCommand.java
@@ -21,6 +21,8 @@ import java.util.List;
 import java.util.concurrent.Callable;
 import javax.inject.Inject;
 
+import io.fabric8.forge.camel.commands.project.helper.DockerSetupHelper;
+import io.fabric8.forge.camel.commands.project.validator.ClassNameValidator;
 import org.jboss.forge.addon.maven.projects.MavenFacet;
 import org.jboss.forge.addon.projects.Project;
 import org.jboss.forge.addon.ui.context.UIBuilder;
