+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/JubeStepCommand.java
@@ -20,6 +20,9 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.inject.Inject;
 
+import io.fabric8.forge.camel.commands.project.helper.DockerSetupHelper;
+import io.fabric8.forge.camel.commands.project.helper.JubeSetupHelper;
+import io.fabric8.forge.camel.commands.project.validator.ClassNameValidator;
 import org.jboss.forge.addon.maven.projects.MavenFacet;
 import org.jboss.forge.addon.projects.Project;
 import org.jboss.forge.addon.projects.dependencies.DependencyInstaller;
