+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/archetype/CamelArchetypeCatalogFactory.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.forge.camel.commands.project;
+package io.fabric8.forge.camel.commands.project.archetype;
 
 import java.io.StringReader;
 import java.util.logging.Level;
