+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/completer/RouteBuilderCompleter.java
@@ -13,7 +13,7 @@
  *  implied.  See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package io.fabric8.forge.camel.commands.project;
+package io.fabric8.forge.camel.commands.project.completer;
 
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
