+++ forge/addons/camel/src/main/java/io/fabric8/forge/camel/commands/project/helper/JubeSetupHelper.java
@@ -13,8 +13,9 @@
  *  implied.  See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package io.fabric8.forge.camel.commands.project;
+package io.fabric8.forge.camel.commands.project.helper;
 
+import io.fabric8.forge.camel.commands.project.helper.VersionHelper;
 import org.jboss.forge.addon.dependencies.Coordinate;
 import org.jboss.forge.addon.dependencies.Dependency;
 import org.jboss.forge.addon.dependencies.builder.CoordinateBuilder;
